requirements.txt
ffmpeg-python==0.2.0
numpy==1.22.2
opencv_contrib_python==4.6.0.66
termcolor==1.1.0
